requirements.txt
djangorestframework==3.14.0
environs==11.0.0
pip-chill==1.0.3
